fix:Combining webview modals into one #17
